added changes on productcontroller

+++ app/Http/Controllers/productController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Category;
 use App\Product;
+use App\User;
 use Illuminate\Http\Request;
 
 class productController extends Controller
@@ -31,7 +32,8 @@ class productController extends Controller
                 'status' => 'required|string',
                 'price' => 'required|numeric',
                 'description' => 'string',
-                'category' => 'required|integer|in:' . implode(',', $categoriesIds)
+                'category_id' => 'required|integer|in:' . implode(',', $categoriesIds),
+                'user_id' => 'required|integer'
 
             ]);
             $product = Product::create($body);
@@ -43,7 +45,7 @@ class productController extends Controller
             ], 500);
         }
     }
-    public function update(Request $request, $id)
+    public function update(Request $request, $product_id, $user_id)
     {
         try {
             $categories = Category::all();
@@ -58,7 +60,11 @@ class productController extends Controller
                 'description' => 'string',
                 'category_id' => 'integer|in:' . implode(',', $categoriesIds)
             ]);
-            $product = Product::find($id);
+            $product = Product::find($product_id);
+            $user = User::find($user_id);
+            if ($product['user_id'] !== $user['id'] or $user['role'] !== 'admin') {
+                return response($message = 'no está autorizado para ejecutar esta accion', 500);
+            }
             $product->update($body);
             return response($product->load('category'));
         } catch (\Exception $e) {
@@ -67,10 +73,14 @@ class productController extends Controller
             ], 500);
         }
     }
-    public function delete($id)
+    public function delete($product_id, $user_id)
     {
         try {
-            $product = Product::find($id);
+            $product = Product::find($product_id);
+            $user = User::find($user_id);
+            if ($user['role'] !== 'admin') {
+                return response($message = 'no está autorizado para ejecutar esta accion', 500);
+            }
             $product->delete();
             return response([
                 'message' => 'Producto eliminado con éxito',
